+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0-rc.11] - 2023-05-16
 ### Added
 - `@distributedlab/w3p`
   - EVM fallback provider
@@ -199,7 +199,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [old repo]: https://github.com/distributed-lab/web-kit-old
 
-[Unreleased]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.10...HEAD
+[Unreleased]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.11...HEAD
+[0.2.0-rc.11]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.10...0.2.0-rc.11
 [0.2.0-rc.10]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.9...0.2.0-rc.10
 [0.2.0-rc.9]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.8...0.2.0-rc.9
 [0.2.0-rc.8]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.7...0.2.0-rc.8